Optimizations

+++ src/camera.rs
@@ -1,6 +1,6 @@
-use std::f64::consts::TAU;
+use std::f32::consts::TAU;
 
-use crate::{ray::Ray, Vector};
+use crate::{ray::Ray, vector::Vector};
 
 pub struct Camera {
     origin: Vector,
@@ -9,7 +9,7 @@ pub struct Camera {
     vertical: Vector,
     u: Vector,
     v: Vector,
-    lens_radius: f64,
+    lens_radius: f32,
 }
 
 impl Camera {
@@ -17,10 +17,10 @@ impl Camera {
         lookfrom: Vector,
         lookat: Vector,
         vup: Vector,
-        vfov: f64,
-        aspect_ratio: f64,
-        aperture: f64,
-        focus_dist: f64,
+        vfov: f32,
+        aspect_ratio: f32,
+        aperture: f32,
+        focus_dist: f32,
     ) -> Self {
         let theta = vfov.to_radians();
         let h = (theta / 2.0).tan();
@@ -46,7 +46,7 @@ impl Camera {
         }
     }
 
-    pub fn get_ray(&self, s: f64, t: f64) -> Ray {
+    pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd = self.lens_radius * random_in_unit_disk();
         let offset = self.u * rd.x() + self.v * rd.y();
 
@@ -63,5 +63,5 @@ fn random_in_unit_disk() -> Vector {
 
     let theta = rng.gen_range(0.0..TAU);
 
-    Vector::from_xyz(theta.sin(), theta.cos(), 0.0) * rng.gen::<f64>()
+    Vector::from_xyz(theta.sin(), theta.cos(), 0.0) * rng.gen::<f32>()
 }
