+++ src/material/dialectric.rs
@@ -5,7 +5,7 @@ use crate::{color::Color, geometry::Hit, ray::Ray};
 use super::{reflect, refract, Material};
 
 pub struct Dialectric {
-    pub index: f64,
+    pub index: f32,
 }
 
 impl Material for Dialectric {
@@ -35,7 +35,7 @@ impl Material for Dialectric {
     }
 }
 
-fn reflectance(cosine: f64, ref_idx: f64) -> f64 {
+fn reflectance(cosine: f32, ref_idx: f32) -> f32 {
     let r0 = (1.0 - ref_idx) / (1.0 + ref_idx);
     let r0 = r0 * r0;
     r0 + (1.0 - r0) * (1.0 - cosine).powi(5)
